test: move test to metrics crate
